+++ src/camera/TopCameraBehave.cpp
@@ -7,7 +7,7 @@ TopCameraBehave::TopCameraBehave(): CameraBehave(Urho3D::Vector3(0, 50, 0), 20, 
 }
 
 void TopCameraBehave::translate(bool cameraKeys[], int wheel, float timeStep, float min) {
-	translateInternal(cameraKeys, timeStep, sqrt((orthoSize - minY) / 10) + 1);
+	translateInternal(cameraKeys, timeStep, sqrt((orthoSize - minY) / 10.f) + 1.f);
 	if (wheel != 0) {
 		orthoSize -= wheel * (sqrt(orthoSize - minY) + 1) * 1.5f;
 
@@ -22,7 +22,7 @@ void TopCameraBehave::translate(bool cameraKeys[], int wheel, float timeStep, fl
 	}
 }
 
-void TopCameraBehave::rotate(const Urho3D::IntVector2& mouseMove, const double mouse_sensitivity) {
+void TopCameraBehave::rotate(const Urho3D::IntVector2& mouseMove, const float mouse_sensitivity) {
 }
 
 void TopCameraBehave::setRotation(const Urho3D::Quaternion& rotation) {
